add chain_selector mod
