+++ crates/floresta-chain/src/pruned_utreexo/chainparams.rs
@@ -12,6 +12,8 @@ use bitcoin::blockdata::constants::genesis_block;
 use bitcoin::Block;
 use bitcoin::BlockHash;
 use bitcoin::Target;
+use rustreexo::accumulator::node_hash::NodeHash;
+use rustreexo::accumulator::stump::Stump;
 
 use crate::prelude::*;
 use crate::Network;
@@ -49,6 +51,8 @@ pub struct ChainParams {
     /// A list of exceptions to the rules, where the key is the block hash and the value is the
     /// verification flags
     pub exceptions: HashMap<BlockHash, c_uint>,
+    pub network_roots: Stump,
+    pub assumeutreexo_index: (BlockHash, u32),
 }
 
 impl ChainParams {
@@ -59,6 +63,32 @@ impl ChainParams {
             Network::Signet => Target::MAX_ATTAINABLE_SIGNET,
             Network::Regtest => Target::MAX_ATTAINABLE_REGTEST,
         }
+    }
+}
+
+fn get_signet_roots() -> Stump {
+    let roots: Vec<NodeHash> = [
+        "8e6fcdcf05020fa1f7131a59a7050b33ca74852f5e82a5fbe236402bc4c8a928",
+        "f4c92949c71be7613699977eebf6d3bd5c8fd3e538a01380583e5aba14273425",
+        "d73ceb2748d342b14a269d7c0feb34aca1341a6367cc75cff6db8422eb01916d",
+        "a012e516784ccb7af26d7b356bf645e6a167cce5b48b9368c58c523acd25f6bf",
+        "e6e74ebc1d01ac47541c90afaac208c9b0f16226d2d046742032374e925a79ae",
+        "235b255558e994e6c5b6011469e891436cbf18107a939847e6e5df4cb939a96b",
+        "a9f45482564f0cb103067636c39fe30df1fa04b6b04d438c655530d991432761",
+        "d46716b7ccaf8d9eff11557527056f6100e016126df369eef95b9c9874467d40",
+        "7039b9053ef819d35c079eb4dcdd37029653a325bf416768e7de16bacf2c90af",
+        "f7a626339303030fc1b71d228e74aebdc2126cb7a2c5e01eb036225ea9dd41c2",
+        "b21123705cb4cef5a104705037ccd80ae7281789aa07cd468d5949c7e62df37b",
+        "ca931559f3ad9c91b9510f5dbfa42467e40ad8a0069d8f273de6079e9b115232",
+        "954ca698b58b6e6cdcc89948c841059d892578b7d67a249965fff83de5aaa7e3",
+    ]
+    .iter()
+    .map(|hash| NodeHash::from_str(hash).unwrap())
+    .collect();
+
+    Stump {
+        roots,
+        leaves: 1477499,
     }
 }
 
@@ -89,6 +119,7 @@ fn get_exceptions() -> HashMap<BlockHash, c_uint> {
 fn get_exceptions() -> HashMap<BlockHash, c_uint> {
     HashMap::new()
 }
+
 impl From<Network> for ChainParams {
     fn from(net: Network) -> Self {
         let genesis = genesis_block(net.into());
@@ -96,6 +127,7 @@ impl From<Network> for ChainParams {
         let exceptions = get_exceptions();
         match net {
             Network::Bitcoin => ChainParams {
+                assumeutreexo_index: (genesis.block_hash(), 0),
                 genesis,
                 max_target,
                 pow_allow_min_diff: false,
@@ -110,8 +142,10 @@ impl From<Network> for ChainParams {
                 segwit_activation_height: 481824,
                 csv_activation_height: 419328,
                 exceptions,
+                network_roots: Stump::default(),
             },
             Network::Testnet => ChainParams {
+                assumeutreexo_index: (genesis.block_hash(), 0),
                 genesis,
                 max_target,
                 pow_allow_min_diff: true,
@@ -126,6 +160,7 @@ impl From<Network> for ChainParams {
                 segwit_activation_height: 834_624,
                 csv_activation_height: 770_112,
                 exceptions,
+                network_roots: Stump::default(),
             },
             Network::Signet => ChainParams {
                 genesis,
@@ -142,8 +177,17 @@ impl From<Network> for ChainParams {
                 bip66_activation_height: 1,
                 segwit_activation_height: 1,
                 exceptions,
+                network_roots: get_signet_roots(),
+                assumeutreexo_index: (
+                    BlockHash::from_str(
+                        "0000001321625245a27e0be82a640106d019e35e48a024a17df1ceeb9b1f2131",
+                    )
+                    .unwrap(),
+                    74551,
+                ),
             },
             Network::Regtest => ChainParams {
+                assumeutreexo_index: (genesis.block_hash(), 0),
                 genesis,
                 max_target,
                 pow_allow_min_diff: false,
@@ -158,6 +202,7 @@ impl From<Network> for ChainParams {
                 bip66_activation_height: 0,
                 segwit_activation_height: 0,
                 exceptions,
+                network_roots: Stump::default(),
             },
         }
     }
