+++ crates/floresta-wire/src/lib.rs
@@ -24,6 +24,8 @@ pub use p2p_wire::node;
 pub use p2p_wire::node_context;
 #[cfg(not(target_arch = "wasm32"))]
 pub use p2p_wire::node_interface;
+#[cfg(not(target_arch = "wasm32"))]
+pub use p2p_wire::running_node;
 
 /// NodeHooks is a trait that defines the hooks that a node can use to interact with the network
 /// and the blockchain. Every time an event happens, the node will call the corresponding hook.
